third_party/blink/web_tests/virtual/select-parser-relaxation/html5lib/generated/run-tests18-data-expected.txt

CONSOLE ERROR: Uncaught SyntaxError: Unexpected token '<'
CONSOLE ERROR: Uncaught SyntaxError: Unexpected token '<'
CONSOLE ERROR: Uncaught SyntaxError: Unexpected token '<'
CONSOLE ERROR: Uncaught SyntaxError: Unexpected token '<'
../resources/tests18.dat:
5
14
15

Test 5 of 36 in ../resources/tests18.dat failed. Input:
<!doctype html><html><noscript><plaintext></plaintext>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|     <noscript>
|       "<plaintext></plaintext>"
|   <body>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|     <noscript>
|   <body>
|     <plaintext>
|       "</plaintext>"

Test 14 of 36 in ../resources/tests18.dat failed. Input:
<!doctype html><select><plaintext></plaintext>X
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       <plaintext>
|         "</plaintext>X"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       "X"

Test 15 of 36 in ../resources/tests18.dat failed. Input:
<!doctype html><table><select><plaintext>a<caption>b
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       <plaintext>
|         "a<caption>b"
|     <table>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       "a"
|     <table>
|       <caption>
|         "b"
